requirements.txt
backoff==1.11.1 ; python_full_version >= "3.7.0" and python_full_version < "4.0.0"
certifi==2022.9.14 ; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.7" and python_version < "4"
deprecation==2.1.0 ; python_full_version >= "3.7.0" and python_full_version < "4.0.0"
docstring-parser==0.14.1 ; python_full_version >= "3.7.0" and python_version < "4.0"
filetype==1.2.0 ; python_full_version >= "3.7.0" and python_full_version < "4.0.0"
idna==3.4 ; python_version >= "3.7" and python_version < "4"
inflection==0.5.1 ; python_full_version >= "3.7.0" and python_full_version < "4.0.0"
packaging==21.3 ; python_full_version >= "3.7.0" and python_full_version < "4.0.0"
pyparsing==3.0.9 ; python_full_version >= "3.7.0" and python_full_version < "4.0.0"
python-slugify==6.1.2 ; python_full_version >= "3.7.0" and python_full_version < "4.0.0"
requests==2.28.1 ; python_version >= "3.7" and python_version < "4"
text-unidecode==1.3 ; python_full_version >= "3.7.0" and python_full_version < "4.0.0"
typing-extensions==4.3.0 ; python_version >= "3.7" and python_full_version < "4.0.0"
urllib3==1.26.12 ; python_version >= "3.7" and python_version < "4"
